add Discounts component
